Extract Store-related iterators into impl module.
We have a bunch of iteration-related logic for scanning over fjall
keyspaces which was previously all defined in the huge store.rs. This
extracts them into a new impl module file shrinking the size of the
store.
